pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "torch>=1.4", "wheel", "ninja"]
build-backend = "setuptools.build_meta"

[project]
name = "eden_grasp_gen"
version = "0.1.2"
description = "A fork of GraspGen for EDEN project"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
  {name = "Kashu Yamazaki", email = "kyamazak@andrew.cmu.edu"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "h5py",
    "hydra-core",
    "matplotlib",
    "meshcat",
    "numpy==1.26.4",
    "webdataset",
    "scikit-learn",
    "scipy",
    "tensorboard",
    "trimesh==4.5.3",
    "transformers",
    "tensordict",
    "diffusers==0.11.1",
    "timm==1.0.15",
    "huggingface-hub==0.25.2",
    "PyOpenGL==3.1.5",
    "addict",
    "spconv-cu126",
    "yapf==0.40.1",
    "tensorboardx",
    "sharedarray",
    "torch-geometric",
    "yourdfpy==0.0.56",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["grasp_gen*"]

[project.scripts]
grasp_gen = "grasp_gen._main:main"

